added multisig option and removed token context
