+++ frontend/src/home/Token/Multisig.js
@@ -1,0 +1,16 @@
+import React from "react";
+
+const Multisig = () => {
+	return (
+		<div className="p-2">
+			<div className="flex justify-center">
+				<h2 className="text-center font-semibold text-[1.4375rem] md:text-[1.75rem]">
+					Multisig
+				</h2>
+			</div>
+			<div className="-mt-1 divider"></div>
+		</div>
+	);
+};
+
+export default Multisig;
