+++ frontend/src/index.js
@@ -4,7 +4,6 @@ import "./index.css";
 import Paths from "./Paths";
 import * as Sentry from "@sentry/react";
 import { BrowserTracing } from "@sentry/tracing";
-import TokenProvider from "./home/Token/TokenProvider";
 
 Sentry.init({
 	dsn: "https://f4a0daaa0b47406f9e612e3b609e8391@o877317.ingest.sentry.io/6744154",
@@ -15,9 +14,4 @@ Sentry.init({
 	tracesSampleRate: 1.0,
 });
 
-ReactDOM.render(
-	<TokenProvider>
-		<Paths />
-	</TokenProvider>,
-	document.getElementById("root")
-);
+ReactDOM.render(<Paths />, document.getElementById("root"));
